+++ README.md
@@ -4,6 +4,6 @@
 
 `monotime` provides a portable monolithic timer.
 
-- Linux: Uses `clock_gettime(CLOCK_MONOTONIC, ...)`
+- Linux: Uses `runtime.nanotime` from [goarista](https://github.com/aristanetworks/goarista/blob/master/monotime/nanotime.go)
 - macOS: Uses `mach_absolute_time` [docs](https://developer.apple.com/library/content/qa/qa1398/_index.html)
